+++ sydney/constants.py
@@ -1,0 +1,12 @@
+HEADERS = {
+    "Accept": "application/json",
+    "Accept-Encoding": "gzip, deflate",
+    "Accept-Language": "en-US,en;q=0.9",
+    "Content-Type": "application/json",
+    "User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/111.0.0.0 Safari/537.36 Edg/111.0.1661.54",
+}
+
+BING_CREATE_CONVESATION_URL = "https://www.bing.com/turing/conversation/create"
+BING_CHATHUB_URL = "wss://sydney.bing.com/sydney/ChatHub"
+
+DELIMETER = "\x1e"  # Record separator character.
