+++ sydney/sydney.py
@@ -1,42 +1,12 @@
 import json
 import os
-from enum import Enum
 
 import websockets.client as websockets
 from aiohttp import ClientSession
 
-HEADERS = {
-    "Accept": "application/json",
-    "Accept-Encoding": "gzip, deflate",
-    "Accept-Language": "en-US,en;q=0.9",
-    "Content-Type": "application/json",
-    "User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/111.0.0.0 Safari/537.36 Edg/111.0.1661.54",
-}
-
-BING_CREATE_CONVESATION_URL = "https://www.bing.com/turing/conversation/create"
-BING_CHATHUB_URL = "wss://sydney.bing.com/sydney/ChatHub"
-
-RECORD_SEPARATOR = "\x1e"
-
-
-def as_json(message: dict) -> str:
-    """
-    Convert message to JSON, append record separator character at the end.
-    """
-    return json.dumps(message) + RECORD_SEPARATOR
-
-
-class ConversationStyle(Enum):
-    """
-    Bing Chat conversation styles. Supported options are:
-    - `creative` for original and imaginative chat
-    - `balanced` for informative and friendly chat
-    - `precise` for concise and straightforward chat
-    """
-
-    creative = "h3relaxedimg"
-    balanced = "galileo"
-    precise = "h3precise"
+from .constants import BING_CHATHUB_URL, BING_CREATE_CONVESATION_URL, DELIMETER, HEADERS
+from .enums import ConversationStyle
+from .utils import as_json
 
 
 class SydneyClient:
@@ -151,7 +121,7 @@ class SydneyClient:
 
         await self.wss_client.send(as_json(request))
         while True:
-            objects = str(await self.wss_client.recv()).split(RECORD_SEPARATOR)
+            objects = str(await self.wss_client.recv()).split(DELIMETER)
             for obj in objects:
                 if not obj:
                     continue
