+++ sydney/utils.py
@@ -1,0 +1,10 @@
+import json
+
+from .constants import DELIMETER
+
+
+def as_json(message: dict) -> str:
+    """
+    Convert message to JSON, append delimeter character at the end.
+    """
+    return json.dumps(message) + DELIMETER
